backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.30
psycopg[binary,pool]>=3.1.18
alembic>=1.13.0
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
boto3>=1.34.0
python-multipart>=0.0.9
celery[redis]>=5.3.0
redis>=5.0.0
fpdf2>=2.7.0
# geoalchemy2>=0.15.0  # Not needed - using text for location
httpx>=0.27.0
PyJWT>=2.9.0
pwdlib[argon2]>=0.3.0
slowapi>=0.1.9
pytest>=8.0.0
pytest-asyncio>=0.23.0
openpyxl>=3.1.0
Pillow>=10.3.0
PyPDF2>=3.0.0
pytest-cov>=5.0.0
